+++ webdoc/Gemfile
@@ -4,8 +4,6 @@ source "https://rubygems.org"
 
 git_source(:github) {|repo_name| "https://github.com/#{repo_name}" }
 
-# gem "rails"
-
 gem 'sinatra'
 gem 'sinatra-contrib'
