+++ webdoc/controller/static.rb
@@ -1,0 +1,5 @@
+
+get '/style.css' do
+  content_type 'text/css'
+  scss :style
+end
