+++ webdoc/dbhelper/collector.rb
