+++ webdoc/dbhelper/dbutil.rb
