+++ webdoc/dbhelper/dbwrapper.rb
@@ -1,4 +1,3 @@
-
 require 'fileutils'
 require 'sqlite3'
